+++ client/CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Added
 
+* Added client generation for updated workflow runs.
 * Added client generation for connection provider models.
 * Added client generation for updated workflow secrets.
 * Expanded client generation for existing specification.